+++ framework/src/test/java/viewpointtest/quals/Bottom.java
@@ -13,6 +13,6 @@ import java.lang.annotation.Target;
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})
-@SubtypeOf({A.class, B.class, ReceiverDependentQual.class})
+@SubtypeOf({A.class, B.class, ReceiverDependentQual.class, Top.class, PolyVP.class, Lost.class})
 @DefaultFor(TypeUseLocation.LOWER_BOUND)
 public @interface Bottom {}